style: enable consistent code formatting
Code formatting can help new contributors adapt to the coding standard in the project.

The default in LazyVim (which I use) is to have autoformatting on, although I'm not sure what is the exact plugin that configures this.

I looked around for examples of how projects normally configure stylua, and came across the one from the widely used telescope from 2021-07-23:
https://github.com/nvim-telescope/telescope.nvim/pull/1040